requirements.txt
Flask==3.1.0
google-generativeai==1.0.0